+++ changelog.md
@@ -7,6 +7,7 @@
 - Updated dev dependencies.
 - Reverted the more specific package `main` field path.
 - Renamed imports in the test index module.
+- Refactored `extractFiles` tests to use `Object.freeze` with input objects and arrays to ensure input isn’t mutated.
 - Amended the changelog entries for v8.0.0 and v9.0.0.
 
 ## 10.0.0